bcn-arm-readers-main/requirements.txt
attrs==21.4.0
bluepy==1.3.0
coverage==6.3
freezegun==1.1.0
helpers==0.2.0
iniconfig==1.1.1
kafka-python==2.0.2
mock==4.0.3
packaging==21.3
parameterized==0.8.1
pluggy==1.0.0
py==1.11.0
PyBluez==0.23
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-dependency==0.5.1
pytest-network==0.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
timeout-decorator==0.5.0
toml==0.10.2
tomli==2.0.0
